# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "CursorPage_ReservationListItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReservationListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "CursorPage[ReservationListItem]",
        "type": "object"
      },
      "CursorPage_ReviewListItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReviewListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "CursorPage[ReviewListItem]",
        "type": "object"
      },
...
      "ErrorCode": {
        "description": "도메인 에러코드 상수(단일 출처). 라우터/서비스는 ``ErrorCode.X``만 사용한다.",
        "enum": [
          "SLOT_CONFLICT",
          "EMAIL_TAKEN",
          "ROOM_LIMIT_REACHED",
          "ROOM_NOT_FOUND",
          "RESERVATION_NOT_FOUND",
          "NOTIFICATION_NOT_FOUND",
          "ACCOUNT_NOT_FOUND",
          "CANCEL_WINDOW_PASSED",
          "REJECT_WINDOW_PASSED",
          "RESERVATION_NOT_COMPLETED",
          "REVIEW_ALREADY_EXISTS",
          "REVIEW_NOT_FOUND",
          "REVIEW_REPLY_FORBIDDEN",
          "REVIEW_REPLY_ALREADY_EXISTS",
          "FORBIDDEN_ROLE",
          "UNAUTHENTICATED",
          "VALIDATION_ERROR",
          "GEOCODING_UNAVAILABLE",
          "LLM_PROVIDER_UNAVAILABLE"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorDetail": {
        "description": "표준 에러 본문의 ``detail`` 객체(OpenAPI 계약 노출용).",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "표준 에러 응답. 후속 라우터가 ``responses={409: {\"model\": ErrorResponse}}``로 참조.",
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "ReviewCreateRequest": {
        "description": "후기 작성 요청(POST 본문) — 별점 1~5 + 텍스트 1~500자(AC1·AC2).\n\n별점 범위 밖·텍스트 길이 위반은 Pydantic이 **선차단**해 422로 되돌린다 — service가 별점/길이를\n재검증하다 ``ValueError``를 500으로 흘리지 않도록 라우터 도달 전에 막는다(반복 함정 #1).\n텍스트는 **필수**(KTH 결정 2 — 빈 후기 금지): ``min_length=1`` + 공백만 방어(strip 후 재검증).\n\n``reservation_id``는 **경로 파라미터**(중첩 라우트)라 본문에 두지 않는다. ``booker_id``는 인증\nprincipal에서 도출하므로 본문에 없다(클라가 작성자를 지정할 수 없음 — RBAC).",
        "properties": {
          "rating": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Rating",
            "type": "integer"
          },
          "text": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "rating",
          "text"
        ],
        "title": "ReviewCreateRequest",
        "type": "object"
      },
...
      "ReviewPublic": {
        "description": "후기 작성 성공 응답(생성된 후기 리소스, AC1).\n\n생성된 ``Review``의 식별·별점·텍스트·작성시각 + 귀속(room_id·reservation_id)을 싣는다.\n작성 주체 본인 응답이라 식별 필드를 노출해도 무방하나, 일관성을 위해 별점/텍스트/시각/귀속만\n싣는다(``booker_id`` 미노출). datetime은 ``...Z``(``isoformat_utc`` — reservations 선례).",
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "rating": {
            "title": "Rating",
            "type": "integer"
          },
          "reservation_id": {
            "format": "uuid",
            "title": "Reservation Id",
            "type": "string"
          },
          "room_id": {
            "format": "uuid",
            "title": "Room Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "reservation_id",
          "room_id",
          "rating",
          "text",
          "created_at"
        ],
        "title": "ReviewPublic",
        "type": "object"
      },
      "ReviewReplyCreateRequest": {
        "description": "제공자 답글 작성 요청(POST 본문) — 텍스트 1~500자(Story 5.6 AC1·AC2).\n\n텍스트 길이/공백 위반은 Pydantic이 **선차단**해 422로 되돌린다 — service가 길이를 재검증하다\n``ValueError``를 500으로 흘리지 않도록 라우터 도달 전에 막는다(반복 함정 #1 —\n``ReviewCreateRequest`` 미러). ``review_id``는 **경로 파라미터**라 본문에 없다.\n``provider_id``는 인증 principal에서 도출하므로 본문에 없다(클라가 작성자 지정 불가 — RBAC).",
        "properties": {
          "text": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "ReviewReplyCreateRequest",
        "type": "object"
      },
      "ReviewReplyPublic": {
        "description": "답글 작성 성공 응답(생성된 답글 리소스, Story 5.6 AC1).\n\n생성된 ``ReviewReply``의 식별·귀속(review_id)·텍스트·작성시각을 싣는다. provider 본인 응답이라\n식별 노출이 무방하나 일관성을 위해 ``provider_id``는 싣지 않는다. datetime ``...Z``.",
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "review_id": {
            "format": "uuid",
            "title": "Review Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "review_id",
          "text",
          "created_at"
        ],
        "title": "ReviewReplyPublic",
        "type": "object"
      },
...
      "RoomCreateRequest": {
        "description": "공간 등록 요청(provider 전용 쓰기, AC1).\n\n범위·enum 위반은 Pydantic 검증 → 1.5 핸들러가 422로 단일화(P3 1차 차단 + 2.1 defer 회수).\n좌표(``lat``/``lng``)·``admin_dong_code``는 클라이언트가 ``GET /rooms/geocode`` 결과에서\n선택해 실어 보낸다(MVP client-trusted — 인증된 provider + DB CHECK 방어, Dev Notes 참조).",
        "properties": {
          "address": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "admin_dong_code": {
            "maxLength": 20,
            "minLength": 1,
            "title": "Admin Dong Code",
            "type": "string"
          },
          "amenities": {
            "items": {
              "enum": [
                "parking",
                "projector_tv",
                "coffee",
                "whiteboard",
                "wifi",
                "etc"
              ],
              "type": "string"
            },
            "title": "Amenities",
            "type": "array"
          },
          "business_hours": {
            "items": {
              "$ref": "#/components/schemas/BusinessHoursInput"
            },
            "minItems": 1,
            "title": "Business Hours",
            "type": "array"
          },
          "capacity": {
            "minimum": 1.0,
            "title": "Capacity",
            "type": "integer"
          },
          "lat": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat",
            "type": "number"
          },
          "lng": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lng",
            "type": "number"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "price_per_hour": {
            "minimum": 0.0,
            "title": "Price Per Hour",
            "type": "integer"
          },
          "room_type": {
            "enum": [
              "open",
              "private"
            ],
            "title": "Room Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "price_per_hour",
          "capacity",
          "room_type",
          "lat",
          "lng",
          "admin_dong_code",
          "business_hours"
        ],
        "title": "RoomCreateRequest",
        "type": "object"
      },
...
      "RoomMapItem": {
        "description": "룸 핀 렌더용 최소 메타 — 첫 진입 지도 좌표 공급 (Story 3.2, AC1·AC2).\n\n각 항목은 한 활성 룸의 **핀을 찍는 데 필요한 최소 정보**(좌표·이름)다. 프론트가 이\n목록을 ``GET /rooms/availability``(``RoomAvailability``)와 ``room_id``로 인메모리\n조인해 핀 좌표(이 응답)에 가용성 색(저쪽 응답)을 입힌다.\n\n**이 네 필드만** 둔다 — 가격·영업시간·부대시설·수용·주소·``provider_id``·\n``created_at``은 **싣지 않는다**. 상세는 바텀시트(Story 3.3)의 신선 단일 조회 책임이고,\n공개 엔드포인트라 ``provider_id`` 노출을 피하며, 과조회를 막는다(L155 신선도 정책).\n\n**키 이름 = ``room_id``**(``RoomAvailability.room_id``와 동일) → 두 공개 GET을\n``room_id``로 조인하기 위함이다. ``Room.id``를 ``room_id``로 노출하므로\n``from_attributes`` 자동 매핑에 의존하지 않고 서비스에서 명시 생성한다\n(``RoomMapItem(room_id=r.id, ...)`` — ``RoomAvailability`` 명시 생성 선례).\n\n**와이어 규약:** snake_case 그대로(``room_id``·``name``·``lat``·``lng``). 시각 필드가\n없어 ``field_serializer``가 불필요하다.",
        "properties": {
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lng": {
            "title": "Lng",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "room_id": {
            "format": "uuid",
            "title": "Room Id",
            "type": "string"
          }
        },
        "required": [
          "room_id",
          "name",
          "lat",
          "lng"
        ],
        "title": "RoomMapItem",
        "type": "object"
      },
      "RoomPublic": {
        "description": "공간 등록 성공 응답(룸 리소스). ``created_at``은 ``...Z``(1.7 ``UserPublic`` 패턴).",
        "properties": {
          "admin_dong_code": {
            "title": "Admin Dong Code",
            "type": "string"
          },
          "amenities": {
            "items": {
              "type": "string"
            },
            "title": "Amenities",
            "type": "array"
          },
          "capacity": {
            "title": "Capacity",
            "type": "integer"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lng": {
            "title": "Lng",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "price_per_hour": {
            "title": "Price Per Hour",
            "type": "integer"
          },
          "provider_id": {
            "format": "uuid",
            "title": "Provider Id",
            "type": "string"
          },
          "room_type": {
            "title": "Room Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "provider_id",
          "name",
          "price_per_hour",
          "capacity",
          "room_type",
          "amenities",
          "lat",
          "lng",
          "admin_dong_code",
          "is_active",
          "created_at"
        ],
        "title": "RoomPublic",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "DeskNow 백엔드 API (Story 1.2 골격).",
    "title": "DeskNow API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/reservations": {
      "get": {
        "description": "예약자 본인의 예약을 최근순 **한 페이지**로 반환한다 → 200(booker, AC1·AC2·AC4 + F 페이징).\n\n본인(``booker_id == principal.user_id``) 예약만 상태 무관(confirmed/cancelled/rejected)으로\n``created_at`` 내림차순 keyset 페이징한다(``CursorPage`` 봉투 — ``items`` + ``next_cursor``;\n다가오는/지난 분류·6h 취소 가능 판정은 FE가 ``slot_starts`` 스냅샷으로 계산 — 4.7 범위 결정 #3).\n``limit``(기본 20·최대 100)·``cursor``(이전 페이지 토큰)는 쿼리. 손상 커서는 422. 미인증=401,\nprovider/admin=403(``_require_booker``). 예약이 없으면 ``items=[]``·``next_cursor=None``(정상 200).\n\n**룸 이름·is_active 합성은 라우터에서**(``session.get(Room, …)``): reservations.service가\nrooms를 import하면 4.9 ``rooms.service → reservations.service``(차감) 역방향과 **순환 import**가\n되므로, service는 ``reservations`` 테이블만 만지고 룸 메타 합성은 조합 계층인 라우터가\n한다(``rooms.models`` import는 router 한정 = service↔service 순환 아님 — favorites\n``build_favorite_item`` 선례). 슬롯 시간은 ``Reservation.slot_starts`` **스냅샷**(취소/거절\n후에도 잔존 — 범위 결정 #1)을 싣는다.\n\n**``has_review`` 합성(Story 5.5 — 후기 작성 게이팅):** ``reviews.service``의\n``reservation_ids_with_review``를 **단일 호출**(N+1 금지)해 본인 후기 존재 예약 집합을 구하고\n각 항목에 ``has_review = reservation.id in ids``를 싣는다. reservations.router →\nreviews.service **단방향**(라우터=조합 계층, room_name 합성과 동형 — service↔service 순환 아님:\nreviews.service는 reservations를 import하지 않는다).",
        "operationId": "reservations_list_reservations",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPage_ReservationListItem_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "List Reservations",
        "tags": [
          "reservations"
        ]
      }
    },
    "/api/v1/reservations/{reservation_id}/reviews": {
      "post": {
        "description": "이용 완료한 본인 예약에 별점+텍스트 후기를 작성한다 → 201 + ReviewPublic(booker, AC1~AC3).\n\n① **소유권 가드(AC3):** ``session.get(Reservation, id)`` → 미존재이거나 ``booker_id``가\n요청자가 아니면 **404 ``RESERVATION_NOT_FOUND``**. 타인 예약을 403이 아니라 404로 막아 **존재\n여부를 누설하지 않는다**(reservations cancel 패턴 미러). 미인증=401·provider/admin=403은\n``_require_booker``. ② **자격·중복은 service:** 이용 완료 아님(취소/거절/미완료) → 409\n``RESERVATION_NOT_COMPLETED``, 이미 작성 → 409 ``REVIEW_ALREADY_EXISTS``. 입력 형식(별점 범위·\n텍스트 길이)은 ``ReviewCreateRequest``가 선차단(422). room_id는 예약에서 도출한다(경로 불요).",
        "operationId": "reviews_create_review",
        "parameters": [
          {
            "in": "path",
            "name": "reservation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Reservation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Create Review",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/reviews/{review_id}/reply": {
      "post": {
        "description": "제공자가 자기 룸 후기에 답글을 단다 → 201 + ReviewReplyPublic(provider, 5.6 AC1·AC2·AC3·AC4).\n\n① **후기 존재(AC2):** ``session.get(Review, id)`` → 미존재면 **404 ``REVIEW_NOT_FOUND``**.\n② **소유권(AC2):** 후기의 룸(``session.get(Room, review.room_id)``)이 요청자 소유가 아니면\n**403 ``REVIEW_REPLY_FORBIDDEN``** — 후기·룸은 공개 데이터라 존재 누설이 무관하고 epic AC가\n명시 403을 요구한다(booker 도메인의 404 비노설과 의도적 분기). 라우터=조합 계층이 ``Room``을\n조회한다(reviews.service는 ``Room``을 만지지 않음 — 순환 금지, 반복 함정 #3). ③ **중복은\nservice:** 이미 답글이 달린 후기 → 409 ``REVIEW_REPLY_ALREADY_EXISTS``. 미인증=401·\nbooker/admin=403 ``FORBIDDEN_ROLE``은 ``_require_provider``. 텍스트 형식은\n``ReviewReplyCreateRequest`` 선차단(422). room_id는 후기에서 도출(경로 불요). 비활성\n룸이라도 본인 룸이면 허용(자기 데이터).",
        "operationId": "reviews_create_reply",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Review Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewReplyCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewReplyPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Create Reply",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/rooms": {
      "get": {
        "description": "활성 룸의 핀 메타(``{room_id, name, lat, lng}``)를 반환한다 → 200(공개, AC1·AC2).\n\n첫 진입 지도 핀 좌표 공급(FR-4). **인증 없음**(탐색 첫 화면 핀은 비로그인 — PRD §FR-2,\n``/availability``와 동일 근거). 가용성(핀 색)은 ``/availability``, 룸 상세는 바텀시트\n(Story 3.3)가 책임진다 — 이 응답은 핀 좌표·이름만 싣는다(과조회 방지). 컬렉션 루트\n``GET \"\"``는 ``POST \"\"``(등록·메서드 다름)·``GET /{room_id}``(동적·경로 세그먼트 다름)와\n충돌하지 않으나, 일관성을 위해 정적 그룹 근처·``PATCH /{room_id}`` 앞에 선언한다.\noperationId = ``rooms_list_rooms``(1.9 규약, 도메인 내 유일).",
        "operationId": "rooms_list_rooms",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoomMapItem"
                  },
                  "title": "Response Rooms List Rooms",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Rooms",
        "tags": [
          "rooms"
        ]
      },
      "post": {
        "description": "공간을 등록한다 → 201 + RoomPublic(provider 전용, AC1·AC5).\n\n인증된 provider의 ``user_id``를 ``provider_id``로 ``create_room`` 호출. 제공자당 1개\n초과는 409, 자정 넘김 영업시간 등 검증 실패는 422(Pydantic→1.5 핸들러).",
        "operationId": "rooms_create_room",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Create Room",
        "tags": [
          "rooms"
        ]
      }
    },
...
    "/api/v1/rooms/{room_id}/reviews": {
      "get": {
        "description": "룸 상세의 후기를 최신순 **한 페이지**로 반환한다 → 200(공개·무인증, AC4 + F 페이징).\n\n룸 상세는 공개 표면(``GET /rooms/{id}`` 동일 근거 — PRD §FR-2)이라 **인증 없음**. 각 후기는\n**별점·텍스트·작성일만**(작성자 식별 미노출 — KTH 결정 1 익명). ``created_at`` keyset 페이징\n(``CursorPage`` 봉투). ``limit``(기본 20·최대 100)·``cursor``는 쿼리, 손상 커서는 422. 후기가\n없으면 ``items=[]``·``next_cursor=None``(정상 200 — 막다른 화면 금지는 FE 빈 상태 카피).\n미존재/비활성 룸도 빈/기존 목록을 낸다(후기는 공개 데이터, 상세 404와 분리).",
        "operationId": "reviews_list_room_reviews",
        "parameters": [
          {
            "in": "path",
            "name": "room_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Room Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPage_ReviewListItem_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "List Room Reviews",
        "tags": [
          "reviews"
        ]
      }
    },
...
  }
}